backend/requirements.txt
flask
flask-cors
sqlalchemy
psycopg2-binary
python-dotenv
requests
simpy
flask-migrate
